Enable Entity Framework core

+++ JOIN.WASM/Shared/Contact.cs
@@ -8,7 +8,7 @@ namespace JOIN.WASM.Shared
 {
     public class Contact
     {
-        public int ContactId { get; set; }
+        public long ContactId { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
